+++ backend/eatandmeet/api/admin.py
@@ -1,7 +1,7 @@
 from django.contrib import admin
 
-from .models import Event
+from .models import Event, Comment
 
 
 admin.site.register(Event)
-
+admin.site.register(Comment)